Refactor Trigger Input Dialog Module Structure
