+++ internal-packages/workflow-designer-ui/src/header/ui/trigger-input-dialog/helpers.ts
@@ -1,10 +1,14 @@
 import type {
 	FlowTrigger,
+	Generation,
 	Input,
 	ParameterItem,
 	TriggerNode,
+	WorkspaceId,
 } from "@giselle-sdk/data-type";
 import type { githubTriggers } from "@giselle-sdk/flow";
+import type { useGenerationRunnerSystem } from "@giselle-sdk/giselle-engine/react";
+import type { buildWorkflowFromNode } from "@giselle-sdk/workflow-utils";
 import type { z } from "zod";
 
 export function buttonLabel(node: TriggerNode) {
@@ -201,4 +205,34 @@ export function toParameterItems(
 		}
 	}
 	return items;
+}
+
+export function createGenerationsForFlow(
+	flow: NonNullable<ReturnType<typeof buildWorkflowFromNode>>,
+	inputs: FormInput[],
+	values: Record<string, string | number>,
+	createGeneration: ReturnType<
+		typeof useGenerationRunnerSystem
+	>["createGeneration"],
+	workspaceId: WorkspaceId,
+) {
+	const generations: Generation[] = [];
+	for (const job of flow.jobs) {
+		for (const operation of job.operations) {
+			const parameterItems =
+				operation.node.content.type === "trigger"
+					? toParameterItems(inputs, values)
+					: [];
+			const generation = createGeneration({
+				origin: { type: "workspace", id: workspaceId },
+				inputs:
+					parameterItems.length > 0
+						? [{ type: "parameters", items: parameterItems }]
+						: [],
+				...operation.generationTemplate,
+			});
+			generations.push(generation);
+		}
+	}
+	return generations;
 }
