+++ internal-packages/workflow-designer-ui/src/header/ui/trigger-input-dialog/index.ts
@@ -5,5 +5,4 @@ export {
 	createInputsFromTrigger,
 	parseFormInputs,
 } from "./helpers";
-export { createGenerationsForFlow } from "./workflow";
 export { TriggerInputDialog } from "./dialog";
